backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
websockets==12.0
requests==2.32.3
google-cloud-speech==2.26.0
google-cloud-translate==3.15.5
coqpit==0.0.17
TTS==0.22.0
webrtcvad==2.0.10
numpy>=1.24.3
soundfile==0.12.1